jsp_host: idempotent register, fallible drain, module split
Task 1: make POST /jsp/1/register idempotent for an identical replay.
If the credential authenticates and the publisher is already Registered
with the same identity triple (agent_id, lifecycle_generation,
source_epoch), return 200 without mutating the reducer or delivering a
duplicate observation message. A different epoch still returns 409.

Task 2b: remove the unreachable JspHostError::WorkerPanicked variant.
Drop cannot return an error, so a panicking worker is now logged via
tracing::error! (without leaking credentials or bootstrap contents)
instead of being silently swallowed.

Task 2e: make ObservationDelivery::drain fallible (Result<Vec<_>,
JspHostError>). A poisoned mutex now surfaces as a typed error at the
app_shell.rs consumption point and as an immediate panic in the test
helper, instead of an opaque timeout.

Split jsp_host/mod.rs into mod.rs (transport/registry/auth) and
launch.rs (bootstrap/coordinator/runtime) to stay within the source-size
gate.

(Fixes #522)
